Move files into modules.
